+++ monocle/db/db.py
@@ -19,11 +19,11 @@ import socket
 import time
 
 from elasticsearch.helpers import bulk
+from elasticsearch.helpers import scan
 from elasticsearch import client
 from elasticsearch.exceptions import NotFoundError
 
 from monocle.db import queries
-from monocle import utils
 
 CHANGE_PREFIX = 'monocle.changes.'
 
@@ -143,6 +143,7 @@ class ELmonocleDB:
                 "assignee": {"type": "keyword"},
                 "approval": {"type": "keyword"},
                 "draft": {"type": "boolean"},
+                "self_merged": {"type": "boolean"},
             }
         }
         settings = {'mappings': self.mapping}
@@ -223,19 +224,6 @@ class ELmonocleDB:
         # especially to be able to set the histogram extended_bounds
         if name not in queries.public_queries:
             raise UnknownQueryException("Unknown query: %s" % name)
-        if not args[1].get('gte'):
-            first_created_event = queries._first_created_event(
-                self.es, self.index, *args, **kwargs
-            )
-            if first_created_event:
-                args[1]['gte'] = int(
-                    utils.is8601_to_dt(first_created_event).timestamp() * 1000
-                )
-            else:
-                # There is probably nothing in the db that match the query
-                args[1]['gte'] = None
-        if not args[1].get('lte'):
-            args[1]['lte'] = int(utils.utcnow().timestamp() * 1000)
         return getattr(queries, name)(self.es, self.index, *args, **kwargs)
 
     def get_indices(self):
@@ -243,3 +231,7 @@ class ELmonocleDB:
             ind.replace(self.prefix, '')
             for ind in self.es.indices.get(self.prefix + '*')
         ]
+
+    def iter_index(self):
+        body = {"query": {"match_all": {}}}
+        return scan(self.es, query=body, index=self.index)